+++ app/Models/User.php
@@ -79,6 +79,6 @@ class User extends Authenticatable
      */
     protected function name(): Attribute
     {
-        return Attribute::make(get: fn (): string => trim("{$this->first_name} {$this->last_name}"));
+        return Attribute::make(get: fn (): string => format_person_name($this->first_name, $this->last_name));
     }
 }